Adds API docs to custom element, clean up README
