+++ examples/perspective_headers.html
@@ -114,7 +114,7 @@
 
             styleListener({detail: regularTable}) {
                 for (const td of regularTable.querySelectorAll("td, thead tr:last-child th")) {
-                    const metadata = regularTable.get_meta(td);
+                    const metadata = regularTable.getMeta(td);
                     let type;
                     if (metadata.x >= 0) {
                         const column_path = this._column_paths[metadata.x];
