+++ package.json
@@ -26,6 +26,7 @@
         "build:less": "lessc src/less/material.less dist/css/material.css",
         "build": "npm-run-all -p build:*",
         "clean": "rimraf dist",
+        "docs": "jsdoc2md src/js/index.js --separators --heading-depth 1 > API.md",
         "lint": "eslint src examples/*.html",
         "fix": "yarn lint --fix",
         "test": "jest --noStackTrace",
@@ -61,6 +62,7 @@
         "http-server": "^0.12.3",
         "jest": "^26.0.1",
         "jest-puppeteer": "^4.4.0",
+        "jsdoc-to-markdown": "^6.0.1",
         "less": "^3.9.0",
         "npm-run-all": "^4.1.3",
         "prettier": "^2.0.5",
